+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,11 +1,12 @@
-package kitchenpos.application;
+package kitchenpos.product.application;
+
+import kitchenpos.dao.ProductDao;
+import kitchenpos.product.domain.Product;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.dao.ProductDao;
-import kitchenpos.domain.Product;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class ProductService {
