+++ src/main/java/kitchenpos/product/ui/ProductRestController.java
@@ -1,7 +1,7 @@
-package kitchenpos.ui;
+package kitchenpos.product.ui;
 
-import kitchenpos.application.ProductService;
-import kitchenpos.domain.Product;
+import kitchenpos.product.application.ProductService;
+import kitchenpos.product.domain.Product;
 import kitchenpos.request.ProductCreateRequest;
 import kitchenpos.response.ProductResponse;
 import org.springframework.http.ResponseEntity;
