+++ src/main/java/kitchenpos/response/ProductResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.response;
 
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 
 import java.math.BigDecimal;
