+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -1,6 +1,7 @@
 package kitchenpos.application;
 
-import kitchenpos.domain.Product;
+import kitchenpos.product.application.ProductService;
+import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
