+++ src/test/java/kitchenpos/domain/ProductTest.java
@@ -1,10 +1,12 @@
 package kitchenpos.domain;
 
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-import java.math.BigDecimal;
+import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import java.math.BigDecimal;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class ProductTest {
